requirements.txt
google-api-python-client==1.7.9
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
pandas==0.24.2
phonenumbers==8.10.12
xmltodict==0.12.0
